+++ src/snn/trainer/training_process/GD_rmsprop_process.cpp
@@ -13,7 +13,7 @@
 namespace SNN
 {
 
-	GD_rmsprop_process::GD_rmsprop_process(Perceptron* p_perceptron, double p_step) : Trainig_process(p_perceptron)
+	GD_rmsprop_process::GD_rmsprop_process(Perceptron* p_perceptron, float p_step) : Trainig_process(p_perceptron)
 	{
 		m_step = p_step;
 		m_Eg = 0;
@@ -31,18 +31,18 @@ namespace SNN
 		else
 			m_gradient *= get_single_derivate();
 
-		m_Eg = 0.9*m_Eg + 0.1*m_gradient*m_gradient;
+		m_Eg = 0.9f*m_Eg + 0.1f*m_gradient*m_gradient;
 
 		add_to_precedent(process, m_gradient);
 
-		m_delta = m_step * m_gradient / sqrt(m_Eg + 0.0001);
+		m_delta = m_step * m_gradient / sqrt(m_Eg + 0.0001f);
 	}
 
 	void GD_rmsprop_process::set_default_configuration(trainig_config_t* configuration)
 	{
 		
 		if (configuration->step == UNDEFINED)
-			configuration->step = 0.002;
+			configuration->step = 0.002f;
 	}
 
 } // namespace SNN_trainer
